Cleaning up some code, and making it not crash on the rake task
I was having issues with the rake task crashing when attempting to print the issue id list
on the rake task. The problem was from 95c31636@app/models/reminder_mailer.rb#L29. Additionally
I made the remind? method in the issue patch return true instead of nil in the elsif case.

Finally, I cleaned up the email a little, both in presentation and string interpolation usage.